+++ HabboHotel/Rooms/RoomManager.cs
@@ -39,6 +39,7 @@ public class RoomManager : IRoomManager
     private readonly IRoomItemStateService _roomItemStateService;
     private readonly IRoomItemPlacementApplyService _roomItemPlacementApplyService;
     private readonly IRoomItemTrackingService _roomItemTrackingService;
+    private readonly IRoomRollerApplyService _roomRollerApplyService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
     private readonly IRoomService _roomService;
@@ -71,6 +72,7 @@ public class RoomManager : IRoomManager
         IRoomItemStateService roomItemStateService,
         IRoomItemPlacementApplyService roomItemPlacementApplyService,
         IRoomItemTrackingService roomItemTrackingService,
+        IRoomRollerApplyService roomRollerApplyService,
         IGameClientManager gameClientManager,
         IGroupManager groupManager,
         IRoomService roomService,
@@ -101,6 +103,7 @@ public class RoomManager : IRoomManager
         _roomItemStateService = roomItemStateService;
         _roomItemPlacementApplyService = roomItemPlacementApplyService;
         _roomItemTrackingService = roomItemTrackingService;
+        _roomRollerApplyService = roomRollerApplyService;
         _clientManager = gameClientManager;
         _groupManager = groupManager;
         _roomService = roomService;
@@ -242,7 +245,7 @@ public class RoomManager : IRoomManager
 
     private Room CreateRoomInstance(RoomData data)
     {
-        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService, _roomItemStateService, _roomItemPlacementApplyService, _roomItemTrackingService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
+        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService, _roomItemStateService, _roomItemPlacementApplyService, _roomItemTrackingService, _roomRollerApplyService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
     }
 
     private static void DisposeRoom(Room room) => room.Dispose();